requirements.txt
auth0-python
